Renamed BCrypt_Hasher package to utilities. It will now contain helper classes like Serializer and Hasher
