+++ src/main/java/utilities/Serializer.java
@@ -1,4 +1,4 @@
-package models.utils;
+package utilities;
 
 import java.io.*;
